ml/requirements.txt
torch==2.5.1
torchvision==0.20.1
numpy==2.2.6
scikit-learn==1.6.1
pandas==2.3.3
pyarrow==18.1.0
joblib==1.4.2
pillow==11.3.0
tqdm==4.67.1
pyyaml==6.0.2
pytest==8.4.1
